done fomr login register

+++ src/redux/user/userSlice.jsx
@@ -1,9 +1,11 @@
 import {createSlice} from "@reduxjs/toolkit";
 import { login, logout } from "../../sevives/useService";
+import { Message } from "@mui/icons-material";
 
 
 const initialState = {
-  currentUser: JSON.parse(localStorage.getItem('user'))
+  currentUser: JSON.parse(localStorage.getItem('user')),
+
 }
 
 const userSlice = createSlice({
@@ -11,8 +13,16 @@ const userSlice = createSlice({
   initialState,
   extraReducers : builder =>{
     builder.addCase(login.fulfilled, (state, action)=>{
-      state.currentUser = action.payload;
-      localStorage.setItem('user', JSON.stringify(action.payload))
+      if(action.payload === "User is not exist"){
+        state.currentUser = undefined;
+      } else if(action.payload === "Password is wrong"){
+        state.currentUser = undefined;
+      }
+      else {
+        state.currentUser = action.payload;
+        localStorage.setItem('user', JSON.stringify(action.payload))
+      }
+      
     })
 
     builder.addCase(logout.fulfilled, (state, action)=>{
